Started adding transforms to Entities to change position.
Doesn't actually change anything yet.
Scene is drawing one Cube and one BRCube which might be incompatible for this approach.

Is it an issue the Cube is using std::move to pass the template vertices to its Entity?
Will that change the prototype data for future cubes.

This is where I left things and took a year long break.